requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.1
django-admin-sortable2==2.2.8
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.6.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.2.1
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
setuptools==57.4.0
sqlparse==0.5.3
stripe==12.4.0
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
